move: plan-detail 컴포넌트 폴더 내부 구조 변경 (#164)
