+++ src/app/plan-detail/_components/core/plan-header.tsx
