+++ src/app/plan-detail/_components/core/plan-map.tsx
