+++ src/app/plan-detail/_components/features/card/place-card-sidemenu.tsx
@@ -7,7 +7,6 @@ import BookmarkIcon from '@/components/commons/bookmark-icon';
 import { Button } from '@/components/ui/button';
 import { CategoryType } from '@/types/category.type';
 
-
 const PLACE_IMAGE_SIZE = {
   width: 40,
   height: 40,
@@ -35,7 +34,7 @@ const handleImageLoadError = (e: React.SyntheticEvent<HTMLImageElement>) => {
   target.style.display = 'none';
 };
 
-const PlaceCardCategory = ({
+const PlaceCardSidemenu = ({
   title,
   imageUrl,
   category,
@@ -111,4 +110,4 @@ const PlaceCardCategory = ({
   );
 };
 
-export default PlaceCardCategory;
+export default PlaceCardSidemenu;
