+++ src/app/plan-detail/_components/features/card/place-card.tsx
