+++ src/app/plan-detail/_components/features/modal/schedule-created-modal.tsx
