+++ src/app/plan-detail/_components/features/sidemenu/place-sidemenu.tsx
