+++ src/app/plan-detail/_components/features/sidemenu/search-sidemenu.tsx
@@ -1,84 +1,63 @@
 'use client';
 
 import { motion, AnimatePresence } from 'framer-motion';
-import { useState, useEffect, useRef, useCallback, useMemo } from 'react';
+import { useEffect, useState, useRef, useCallback, useMemo } from 'react';
 
-import ErrorMessage from '@/components/features/alert/error-message';
 import { Button } from '@/components/ui/button';
 import DynamicPagination from '@/components/ui/dynamic-pagination';
+import { Input } from '@/components/ui/input';
+import fetchGetAllPlaces from '@/lib/apis/search/get-place.api';
 import { useBookmarkQuery } from '@/lib/hooks/use-bookmark-query';
 import { CategoryType } from '@/types/category.type';
+import { Place } from '@/types/search.type';
 
-import PlaceCardCategory from './place-card-category';
+import DaySelectRequiredModal from '../modal/day-select-required-modal';
+import PlaceCardSidemenu from '../card/place-card-sidemenu';
 import PlaceSidemenuLayout from './place-sidemenu-layout';
 
 const ITEMS_PER_PAGE = 7;
-const INITIAL_ITEMS = 3;
+const INITIAL_ITEMS = 5;
 const NAVIGATION_BUTTON_WIDTH = 42.4;
 const NAVIGATION_BUTTON_GAP = 4;
 const DEBOUNCE_TIME = 200;
 
-/**
- * 북마크된 장소들을 사이드메뉴에 표시하는 컴포넌트
- *
- * @param userId - 현재 로그인한 사용자의 ID
- * @param filterTabs - 카테고리 필터 탭 목록
- * @param activeFilterTab - 현재 선택된 필터 탭
- * @param onFilterTabChange - 필터 탭 변경 핸들러
- */
-const BookmarkSidemenu = ({
-  userId,
+const SearchSidemenu = ({
   filterTabs,
   activeFilterTab,
   onFilterTabChange,
+  userId,
+  onAddPlace,
+  selectedDay,
 }: {
-  userId: string;
   filterTabs: CategoryType[];
   activeFilterTab: CategoryType;
   onFilterTabChange: (tab: CategoryType) => void;
+  userId: string;
+  onAddPlace: (place: Place) => void;
+  selectedDay: number | null;
 }) => {
+  const [places, setPlaces] = useState<Place[]>([]);
+  const [searchQuery, setSearchQuery] = useState('');
   const [isLoading, setIsLoading] = useState(true);
-  const [error, setError] = useState<string | null>(null);
   const [isExpanded, setIsExpanded] = useState<boolean>(false);
   const [currentPage, setCurrentPage] = useState(1);
   const [maxVisiblePages, setMaxVisiblePages] = useState(3);
+  const [isDaySelectModalOpen, setIsDaySelectModalOpen] = useState(false);
   const containerRef = useRef<HTMLDivElement>(null);
   const resizeTimeoutRef = useRef<NodeJS.Timeout>();
   const [scrollPosition, setScrollPosition] = useState(0);
   const listRef = useRef<HTMLDivElement>(null);
 
-  const { isBookmarked, toggleBookmark, bookmarks } = useBookmarkQuery(userId);
+  const { isBookmarked, toggleBookmark } = useBookmarkQuery(userId);
 
-  const handleBookmarkToggle = async (place_id: number) => {
-    try {
-      // 현재 스크롤 위치 저장
-      if (listRef.current) {
-        setScrollPosition(listRef.current.scrollTop);
-      }
-
-      await toggleBookmark(place_id);
-
-      // 북마크 상태 변경 후 스크롤 위치 복원
-      setTimeout(() => {
-        if (listRef.current) {
-          listRef.current.scrollTop = scrollPosition;
-        }
-      }, 0);
-    } catch {
-      setError('북마크를 업데이트하는 데 실패했습니다.');
-    }
-  };
-
-  useEffect(() => {
-    setIsLoading(false);
-  }, [bookmarks]);
-
-  /* 카테고리별 필터링 */
-  const filteredPlaces = bookmarks
-    ? activeFilterTab === '전체'
-      ? bookmarks
-      : bookmarks.filter((place) => place.category === activeFilterTab)
-    : [];
+  const filteredPlaces = places.filter((place) => {
+    const matchesCategory =
+      activeFilterTab === '전체' || place.category === activeFilterTab;
+    const matchesSearch =
+      place.title.toLowerCase().includes(searchQuery.toLowerCase()) ||
+      place.address.toLowerCase().includes(searchQuery.toLowerCase());
+    return matchesCategory && matchesSearch;
+  });
 
   const totalPages = Math.ceil(filteredPlaces.length / ITEMS_PER_PAGE);
 
@@ -114,7 +93,7 @@ const BookmarkSidemenu = ({
 
     resizeTimeoutRef.current = setTimeout(() => {
       setMaxVisiblePages(calculateMaxVisiblePages());
-    }, DEBOUNCE_TIME);
+    }, DEBOUNCE_TIME); // 200ms 디바운스 -> resize 이벤트가 빈번하게 일어나 계산이 너무 자주 일어나지 않도록 구현
   }, [calculateMaxVisiblePages]);
 
   useEffect(() => {
@@ -130,6 +109,56 @@ const BookmarkSidemenu = ({
       }
     };
   }, [calculateMaxVisiblePages, updateMaxVisiblePages]);
+
+  useEffect(() => {
+    const fetchData = async () => {
+      try {
+        const placesResponse = await fetchGetAllPlaces();
+        setPlaces(placesResponse || []);
+      } catch (error) {
+        console.error('데이터를 가져오는데 실패했습니다:', error);
+      } finally {
+        setIsLoading(false);
+      }
+    };
+
+    fetchData();
+  }, [userId]);
+
+  const handleBookmarkToggle = async (placeId: number) => {
+    try {
+      // 현재 스크롤 위치 저장
+      if (listRef.current) {
+        setScrollPosition(listRef.current.scrollTop);
+      }
+
+      await toggleBookmark(placeId);
+
+      // 북마크 상태 변경 후 스크롤 위치 복원
+      setTimeout(() => {
+        if (listRef.current) {
+          listRef.current.scrollTop = scrollPosition;
+        }
+      }, 0);
+    } catch (error) {
+      console.error('북마크 토글에 실패했습니다:', error);
+    }
+  };
+
+  const handleAddPlace = (place: Place) => {
+    if (selectedDay === null) {
+      setIsDaySelectModalOpen(true);
+      return;
+    }
+
+    // 장소의 위치 정보가 있는지 확인
+    if (!place.lat || !place.lng) {
+      console.error('장소의 위치 정보가 없습니다:', place);
+      return;
+    }
+
+    onAddPlace(place);
+  };
 
   const handlePageChange = (page: number) => {
     setCurrentPage(page);
@@ -150,51 +179,66 @@ const BookmarkSidemenu = ({
     ? currentPageItems
     : currentPageItems.slice(0, INITIAL_ITEMS);
 
-  if (error) {
+  const searchBar = (
+    <div className="rounded-[12px] bg-gray-100 px-3 py-2">
+      <div className="flex items-center gap-[12px]">
+        <img src="/icons/search.svg" alt="검색" className="h-5 w-5" />
+        <Input
+          type="text"
+          placeholder="장소를 검색해 추가하세요"
+          value={searchQuery}
+          onChange={(e) => setSearchQuery(e.target.value)}
+          className="w-full border-none bg-transparent text-14 font-medium leading-[150%] text-[#698EA1] placeholder:text-gray-400 focus:outline-none focus-visible:ring-0 focus-visible:ring-offset-0"
+        />
+      </div>
+    </div>
+  );
+
+  if (isLoading) {
     return (
-      <ErrorMessage
-        title={'북마크 데이터를 불러오는데 실패했습니다.'}
-        description={error}
-      />
+      <PlaceSidemenuLayout
+        isBookmarkSection={false}
+        filterTabs={filterTabs}
+        activeFilterTab={activeFilterTab}
+        onFilterTabChange={onFilterTabChange}
+        topContent={searchBar}
+      >
+        <div className="flex items-center justify-center p-4">
+          <p>로딩 중...</p>
+        </div>
+      </PlaceSidemenuLayout>
     );
   }
 
   return (
     <PlaceSidemenuLayout
-      isBookmarkSection
+      isBookmarkSection={false}
       filterTabs={filterTabs}
       activeFilterTab={activeFilterTab}
       onFilterTabChange={onFilterTabChange}
+      topContent={searchBar}
     >
-      {/* 북마크 리스트 */}
+      {/* 검색 결과 */}
       <div className="space-y-2" ref={containerRef}>
-        {isLoading ? (
-          <div className="flex items-center justify-center py-4 text-gray-500">
-            불러오는 중...
-          </div>
-        ) : filteredPlaces.length === 0 ? (
-          <div className="flex items-center justify-center py-4 text-gray-500">
-            {activeFilterTab === '전체'
-              ? '북마크한 장소 정보가 없습니다.'
-              : `북마크한 ${activeFilterTab} 정보가 없습니다.`}
-          </div>
-        ) : (
+        {displayedPlaces.length > 0 ? (
           <div ref={listRef} className="overflow-y-auto">
             <AnimatePresence initial={false}>
               {displayedPlaces.map((place) => (
                 <motion.div
-                  key={place.placeId}
+                  key={place.id}
                   initial={{ opacity: 0, height: 0 }}
                   animate={{ opacity: 1, height: 'auto' }}
                   exit={{ opacity: 0, height: 0 }}
                   transition={{ duration: 0.2 }}
                 >
-                  <PlaceCardCategory
+                  <PlaceCardSidemenu
                     title={place.title}
                     category={place.category as CategoryType}
-                    imageUrl={place.image}
+                    imageUrl={place.image || ''}
                     isBookmarked={isBookmarked(place.placeId)}
+                    isSearchSection
                     onBookmarkToggle={() => handleBookmarkToggle(place.placeId)}
+                    onAddPlace={() => handleAddPlace(place)}
                   />
                 </motion.div>
               ))}
@@ -216,10 +260,18 @@ const BookmarkSidemenu = ({
               />
             )}
           </div>
+        ) : (
+          <div className="flex items-center justify-center p-4">
+            <p>검색 결과가 없습니다.</p>
+          </div>
         )}
       </div>
+      <DaySelectRequiredModal
+        isOpen={isDaySelectModalOpen}
+        onClose={() => setIsDaySelectModalOpen(false)}
+      />
     </PlaceSidemenuLayout>
   );
 };
 
-export default BookmarkSidemenu;
+export default SearchSidemenu;
